Removed messages package.
The `messages` package has a bug (really, it's just not a complete
implementation) where a user could push a lot of tokens into server
memory and they might never get cleared. The application that I'm using
these packages in could use a much simpler solution, so I'm going to
remove this package.